feat: add responvise page view
